Extract generic theme utilities to @/components/themes
Moves isDarkMode, prefersDarkMode, documentHasDarkMode, and ThemeMode to a
shared location at @/components/themes. The editor/themes.ts re-exports
these for backwards compatibility. This aligns with nteract/elements which
expects theme utilities at @/components/themes.
